+++ include/transport2d.cuh
@@ -4,46 +4,46 @@
 
 class TransportSolver2D {
 public:
-    TransportSolver2D(int nx, int ny, float dx, float dy, float dt, int num_species);
+    TransportSolver2D(int nx, int ny, double dx, double dy, double dt, int num_species);
     ~TransportSolver2D();
-    void solve(std::vector<float>& concentrations);
-    void setVelocity(float vx, float vy);
-    void setDiffusion(float dx, float dy);
+    void solve(std::vector<double>& concentrations);
+    void setVelocity(double vx, double vy);
+    void setDiffusion(double dx, double dy);
     void setMask(const std::vector<int>& mask);
-    void setCellVolumes(const std::vector<float>& volumes); // apply the fluid volume of each cell accroding to the porosity
+    void setCellVolumes(const std::vector<double>& volumes); // apply the fluid volume of each cell accroding to the porosity
     void checkCFLCondition();
     void checkBoundaryFluxes();
-    float getTotalMass() const;
+    double getTotalMass() const;
 
-    std::vector<float> getDiffusionCoefficients() const;
-    void setModifiedDiffusion(const std::vector<float>& modified_diffusion);
+    std::vector<double> getDiffusionCoefficients() const;
+    void setModifiedDiffusion(const std::vector<double>& modified_diffusion);
     std::vector<int> getMask() const;
-    std::vector<float> getCellVolumes() const;
-    std::vector<float> getPorosity() const;
-    void setPorosity(const std::vector<float>& porosity);
+    std::vector<double> getCellVolumes() const;
+    std::vector<double> getPorosity() const;
+    void setPorosity(const std::vector<double>& porosity);
 
-    void setInterfaceConcentration(float co2_interface_conc) {
+    void setInterfaceConcentration(double co2_interface_conc) {
         interface_concentration_ = co2_interface_conc;
     }
 
 private:
     int nx_, ny_;
-    float dx_, dy_;
-    float dt_;
+    double dx_, dy_;
+    double dt_;
     int num_species_;
-    float* d_concentrations_;
-    float* d_concentrations_new_;
-    float* d_fluxes_x_;      // Fluxes at x-interfaces
-    float* d_fluxes_y_;      // Fluxes at y-interfaces
-    float* d_cell_volumes_;
-    float2 velocity_;
-    float2 diffusion_;
+    double* d_concentrations_;
+    double* d_concentrations_new_;
+    double* d_fluxes_x_;      // Fluxes at x-interfaces
+    double* d_fluxes_y_;      // Fluxes at y-interfaces
+    double* d_cell_volumes_;
+    double2 velocity_;
+    double2 diffusion_;
     int* d_mask_;     
-    float* d_modified_diffusion_;  // Add new member for modified diffusion coefficients
+    double* d_modified_diffusion_;  // Add new member for modified diffusion coefficients
     bool has_modified_diffusion_;  // Flag to track if modified diffusion is being used   
     bool has_mask_;
 
-    float interface_concentration_; // Concentration at scCO2-water interface
-    float* d_interface_flux_;      // Store interface flux for each cell
-    float* d_porosity_;
+    double interface_concentration_; // Concentration at scCO2-water interface
+    double* d_interface_flux_;      // Store interface flux for each cell
+    double* d_porosity_;
 };
